Implemented Login & Signup Page (#17)
* wip: Checking in so others can look at code

* wip: Added code for Sign Up.  Got submit to be called.  More work
needed.  This is a checkpoint for Mauro to review.

* Co-authored-by: Mauro Dorigo <purisquiri@users.noreply.github.com>
Co-authored-by: steranka <steranka@users.noreply.github.com>

* wip: Added functionality to login.  The backend is now called and the
user state data is added to the Redux store.
Several issues remain:
(1) If user enters bad password, the form is not "reset" and needs to
be.  At this point, the user can't login if they get there password
wrong because the submit button is disabled.
(2) The layout (css) still needs work aligning the fields.
(3) The way axios handles errors is through promises so try/catch
doesn't work.  Leave it this way so Eric can see the issue.

* Signup Page changed to handle the case when a user already exists.
An HTTP Code 409 is returned from the backened.

* Revert deleted styles

Co-authored-by: Eric Masi <eric.masi90@gmail.com>